Fixed deadlock...
We encountered the following deadlock: Most python function calls here
eventually descend down to the C++ level, and many block on future::get(). While
a function is blocked on future::get(), a broadcast may arrive from the robot,
which may trigger a callback. The first thing the callback does is acquire the
Python GIL; However, since Python is currently descended into C++ land waiting
on future::get(), the callback gets blocked waiting for the GIL. Simultaneously,
the thread executing the callback function is the same thread which must fulfil
future::get(), and so the application gets deadlocked.

We solved this by executing callbacks in a separate c++ thread.